Catch InterruptedException in Fork.joinEither
